+++ ts-warp.c
@@ -54,7 +54,6 @@
 #include <sys/types.h>
 #include <pwd.h>
 
-#include "ts-warp.h"
 #include "network.h"
 #include "utility.h"
 #include "socks.h"
@@ -64,7 +63,7 @@
 #include "pidfile.h"
 #include "pidlist.h"
 #include "natlook.h"
-
+#include "ts-warp.h"
 
 /* ------------------------------------------------------------------------------------------------------------------ */
 int loglevel = LOG_LEVEL_DEFAULT;
@@ -132,7 +131,7 @@ All parameters are optional:
     struct sockaddr_storage caddr;                                      /* Client address */
     socklen_t caddrlen;                                                 /* Client address len */
     struct sockaddr_storage daddr;                                      /* Client destination address */
-    char dname[SOCKS5_ATYPE_NAME_LEN] = {0};                            /* Client destination name */
+    struct uvaddr daddr_u;                                              /* Client destination ip and/or name */
 
     unsigned char auth_method;                                          /* Socks5 accepted auth method */
 
@@ -412,22 +411,27 @@ All parameters are optional:
         printl(LOG_INFO, "Client: [%d], IP: [%s] accepted", cn++, inet2str(&caddr, buf));
 
         /* -- Get the client original destination from NAT ---------------------------------------------------------- */
-        socklen_t daddrlen = sizeof daddr;                          /* Client dest address len */
+        
+        /* Initialize  daddr_u */
+        daddr_u.ip_addrlen = sizeof(daddr_u.ip_addr);
+        memset(&daddr_u.ip_addr, 0, daddr_u.ip_addrlen);
+        memset(&daddr_u.name, 0, sizeof(daddr_u.name) - 1);
+
         #if defined(linux)
             /* On Linux && nftabeles/iptables */
-            memset(&daddr, 0, daddrlen); 
-            daddr.ss_family = caddr.ss_family;
-            ret = getsockopt(csock, SOL_IP, SO_ORIGINAL_DST, &daddr, &daddrlen);
+            memset(&daddr_u.ip_addr, 0, daddr_u.ip_addrlen);
+            daddr_u.ip_addr.ss_family = caddr.ss_family;
+            ret = getsockopt(csock, SOL_IP, SO_ORIGINAL_DST, &daddr_u.ip_addr, &daddr_u.ip_addrlen);
         #else
             /* On *BSD with PF */
-            ret = nat_lookup(pfd, &caddr, (struct sockaddr_storage *)ires->ai_addr, &daddr);
+            ret = nat_lookup(pfd, &caddr, (struct sockaddr_storage *)ires->ai_addr, &daddr_u.ip_addr);
         #endif
         if (ret != 0) {
             printl(LOG_WARN, "Failed to find the real destination IP, trying to get it from the socket");
-            getpeername(csock, (struct sockaddr *)&daddr, &daddrlen);
+            getpeername(csock, (struct sockaddr *)&daddr_u.ip_addr, &daddr_u.ip_addrlen);
         }
 
-        printl(LOG_INFO, "The client destination address is: [%s]", inet2str(&daddr, buf));
+        printl(LOG_INFO, "The client destination address is: [%s]", inet2str(&daddr_u.ip_addr, buf));
 
         /* -- Process the PIDs list: remove exitted clients and execute workload balance functions ------------------ */
         c = pids;
@@ -456,7 +460,7 @@ All parameters are optional:
         }
 
         /* Find Socks server to serve the destination address in INI file */
-        s_ini = ini_look_server(ini_root, daddr, NULL);
+        s_ini = ini_look_server(ini_root, daddr_u);
         if (s_ini && s_ini->section_balance == SECTION_BALANCE_ROUNDROBIN) pushback_ini(&ini_root, s_ini);
 
         if ((cpid = fork()) == -1) {
@@ -481,14 +485,15 @@ All parameters are optional:
 
             if (!s_ini) {
                 /* -- No proxy server found for the destination IP -------------------------------------------------- */
-                printl(LOG_INFO, "No proxy server is defined for the destination: [%s]", inet2str(&daddr, buf));
+                printl(LOG_INFO, "No proxy server is defined for the destination: [%s]",
+                    inet2str(&daddr_u.ip_addr, buf));
 
-                if ((daddr.ss_family == AF_INET && S4_ADDR(daddr) == S4_ADDR(*ires->ai_addr)) ||
-                    (daddr.ss_family == AF_INET6 && 
-                        !memcmp(S6_ADDR(daddr), S6_ADDR(*ires->ai_addr), sizeof(S6_ADDR(daddr)))) ||
-                    (daddr.ss_family == AF_INET && S4_ADDR(daddr) == S4_ADDR(*hres->ai_addr)) ||
-                    (daddr.ss_family == AF_INET6 &&
-                        !memcmp(S6_ADDR(daddr), S6_ADDR(*hres->ai_addr), sizeof(S6_ADDR(daddr))))) {
+                if ((daddr_u.ip_addr.ss_family == AF_INET && S4_ADDR(daddr_u.ip_addr) == S4_ADDR(*ires->ai_addr)) ||
+                    (daddr_u.ip_addr.ss_family == AF_INET6 && 
+                        !memcmp(S6_ADDR(daddr_u.ip_addr), S6_ADDR(*ires->ai_addr), sizeof(S6_ADDR(daddr_u.ip_addr)))) ||
+                    (daddr_u.ip_addr.ss_family == AF_INET && S4_ADDR(daddr_u.ip_addr) == S4_ADDR(*hres->ai_addr)) ||
+                    (daddr_u.ip_addr.ss_family == AF_INET6 &&
+                        !memcmp(S6_ADDR(daddr_u.ip_addr), S6_ADDR(*hres->ai_addr), sizeof(S6_ADDR(daddr_u.ip_addr))))) {
 
                     /* Desination address:port is the same as ts-warp incominig (Socks or HTTP) ip:port, i.e., a client
                     contacted ts-warp dirctly: no NAT/redirection and TS-Warp is not defined as proxy server */
@@ -498,14 +503,16 @@ All parameters are optional:
                 }
 
                 /*  -- Direct connection with the destination address bypassing proxy ------------------------------- */
-                printl(LOG_INFO, "Making a direct connection with the destination address: [%s]", inet2str(&daddr, buf));
-                if ((ssock = connect_desnation(*(struct sockaddr *)&daddr)) == -1) {
-                    printl(LOG_WARN, "Unable to connect with destination: [%s]", inet2str(&daddr, buf));
+                printl(LOG_INFO, "Making direct connection with the destination address: [%s]",
+                    inet2str(&daddr_u.ip_addr, buf));
+
+                if ((ssock = connect_desnation(*(struct sockaddr *)&daddr_u.ip_addr)) == -1) {
+                    printl(LOG_WARN, "Unable to connect with destination: [%s]", inet2str(&daddr_u.ip_addr, buf));
                     close(csock);
                     exit(1);
                 }
 
-                printl(LOG_INFO, "Succesfully connected with desination address: [%s]", inet2str(&daddr, buf));
+                printl(LOG_INFO, "Succesfully connected with desination address: [%s]", inet2str(&daddr_u.ip_addr, buf));
             } else
                 /* -- Internal TS-Warp proxy servers ---------------------------------------------------------------- */
                 if (isock == Ssock) {
@@ -513,9 +520,9 @@ All parameters are optional:
                     close(Ssock);
                     close(Hsock);
 
-                    if ((daddr.ss_family == AF_INET && S4_ADDR(daddr) == S4_ADDR(*ires->ai_addr)) ||
-                        (daddr.ss_family == AF_INET6 && !memcmp(S6_ADDR(daddr), S6_ADDR(*ires->ai_addr),
-                            sizeof(S6_ADDR(daddr))))) {
+                    if ((daddr_u.ip_addr.ss_family == AF_INET && S4_ADDR(daddr_u.ip_addr) == S4_ADDR(*ires->ai_addr)) ||
+                        (daddr_u.ip_addr.ss_family == AF_INET6 && !memcmp(S6_ADDR(daddr_u.ip_addr),
+                            S6_ADDR(*ires->ai_addr), sizeof(S6_ADDR(daddr_u.ip_addr))))) {
 
                         /* Desination address:port is the same as ts-warp income ip:port, i.e., a client contacted 
                         ts-warp directly: no NAT/redirection, but TS-Warp is indicated as Socks-server */
@@ -527,33 +534,18 @@ All parameters are optional:
                             exit(1);
                         }
 
-                        memset(&dname, 0, sizeof(dname) - 1);
-                        memset(&daddr, 0, sizeof(struct sockaddr_storage));
-                        if ((socks5_server_request(csock,(struct sockaddr_storage *)(ires->ai_addr),
-                                &daddr, dname)) != SOCKS5_ATYPE_NONE)
-                            /* Now daddr must contain TCP-port number regardless of the addr_type */
-                            /* Let's find the real destination, the client wants to reach */
-                            s_ini = ini_look_server(ini_root, daddr, dname);
-                        else
-                            s_ini = ini_look_server(ini_root, daddr, NULL);
+                        if ((socks5_server_request(csock, 
+                            (struct sockaddr_storage *)(ires->ai_addr), &daddr_u)) != SOCKS5_ATYPE_NONE)
+                                s_ini = ini_look_server(ini_root, daddr_u);
 
                         if (!s_ini || (s_ini && SA_FAMILY(s_ini->socks_server) == AF_INET ? \
                             S4_ADDR(s_ini->socks_server) == S4_ADDR(*ires->ai_addr) : \
                             S6_ADDR(s_ini->socks_server) == S6_ADDR(*ires->ai_addr))) {
 
                             printl(LOG_INFO, "Serving request to: [%s] with embedded TS-Warp SOCKS server",
-                                dname[0] ? dname : inet2str(&daddr, buf));
+                                daddr_u.name[0] ? daddr_u.name : inet2str(&daddr_u.ip_addr, buf));
 
-                            if (dname[0]) {
-                                /* Convert hostname and port to struct sockaddr_storage */
-                                char *dport;
-                                asprintf(&dport, "%i",
-                                    SA_FAMILY(daddr) == AF_INET ? htons(SIN4_PORT(daddr)) : htons(SIN6_PORT(daddr)));
-                                daddr = str2inet(dname, dport);
-                                free(dport);
-                            }
-
-                            if ((ssock = connect_desnation(*(struct sockaddr *)&daddr)) == -1) {
+                            if ((ssock = connect_desnation(*(struct sockaddr *)&daddr_u.ip_addr)) == -1) {
                                 printl(LOG_WARN, "Unable to connect with destination: [%s]",
                                     inet2str(&daddr, buf));
                                 close(csock);
@@ -564,7 +556,7 @@ All parameters are optional:
 
                         } else
                             printl(LOG_INFO, "Serving request to [%s : %s] with external Socks, section name: [%s]",
-                                dname, inet2str(&daddr, buf), s_ini->section_name);
+                                daddr_u.name, inet2str(&daddr_u.ip_addr, buf), s_ini->section_name);
 
                         /* Pass the client to external Socks-servers - proxy forwarding */
                     }
@@ -575,17 +567,15 @@ All parameters are optional:
                         close(Hsock);
                         /* TODO: Implement HTTP proxy */
 
-                    if ((daddr.ss_family == AF_INET && S4_ADDR(daddr) == S4_ADDR(*ires->ai_addr)) ||
-                        (daddr.ss_family == AF_INET6 && !memcmp(S6_ADDR(daddr), S6_ADDR(*ires->ai_addr),
-                            sizeof(S6_ADDR(daddr))))) {
+                    if ((daddr_u.ip_addr.ss_family == AF_INET && S4_ADDR(daddr_u.ip_addr) == S4_ADDR(*ires->ai_addr)) ||
+                        (daddr_u.ip_addr.ss_family == AF_INET6 && !memcmp(S6_ADDR(daddr_u.ip_addr), S6_ADDR(*ires->ai_addr),
+                            sizeof(S6_ADDR(daddr_u.ip_addr))))) {
 
                         /* Desination address:port is the same as ts-warp income ip:port, i.e., a client contacted 
                         ts-warp directly: no NAT/redirection, but TS-Warp is indicated as HTTP-server */
                         printl(LOG_INFO, "Serving the client with embedded TS-Warp HTTP-server");
 
-                        memset(&dname, 0, sizeof(dname) - 1);
-                        memset(&daddr, 0, sizeof(struct sockaddr_storage));
-                        http_server_request(csock, &daddr, dname);
+                        http_server_request(csock, &daddr_u);       /* TODO: Use daddr_u only! */
 
                         printl(LOG_CRIT, "Internal HTTP proxy is not implemented yet! Dropping the connection");
                         close(csock);
@@ -772,32 +762,35 @@ All parameters are optional:
                     }
 
                     printl(LOG_VERB, "Initiate Socks5 protocol: request [%s] -> [%s]", 
-                        inet2str(&s_ini->socks_server, suf), inet2str(&daddr, buf));
+                        inet2str(&s_ini->socks_server, suf), inet2str(&daddr_u.ip_addr, buf));
 
                     /* -- Perform NIT Lookup ------------------------------------------------------------------------ */
-                    if (daddr.ss_family == AF_INET && S4_ADDR(s_ini->socks_server) != S4_ADDR(daddr)) {
+                    if (daddr_u.ip_addr.ss_family == AF_INET && S4_ADDR(s_ini->socks_server) != S4_ADDR(daddr_u.ip_addr)) {
                         if (s_ini->nit_domain &&
-                            (S4_ADDR(s_ini->nit_ipaddr) & S4_ADDR(s_ini->nit_ipmask)) == (S4_ADDR(daddr) &
+                            (S4_ADDR(s_ini->nit_ipaddr) & S4_ADDR(s_ini->nit_ipmask)) == (S4_ADDR(daddr_u.ip_addr) &
                                 S4_ADDR(s_ini->nit_ipmask)))
-                                    if (getnameinfo((const struct sockaddr *)&daddr, sizeof(daddr), 
-                                        dname, sizeof(dname), 0, 0, NI_NAMEREQD)) {
+                                    if (getnameinfo((const struct sockaddr *)&daddr_u.ip_addr, sizeof(daddr_u.ip_addr), 
+                                        daddr_u.name, sizeof(daddr_u.name), 0, 0, NI_NAMEREQD)) {
                                             printl(LOG_WARN, "Unable to resolve client destination address [%s] via NIT",
-                                                inet2str(&daddr, buf));
+                                                inet2str(&daddr_u.ip_addr, buf));
                                             close(csock);
                                             exit(2);
                                     }
                     }
 
-                    if (socks5_client_request(ssock, SOCKS5_CMD_TCPCONNECT, &daddr, dname) > 0) {    
+                    /* TODO: Use daddr_u only! */
+                    if (socks5_client_request(ssock, SOCKS5_CMD_TCPCONNECT, &daddr_u.ip_addr, daddr_u.name) > 0) {    
                         printl(LOG_CRIT, "Socks server returned an error");
                         close(csock);
                         exit(2);
                     }
                 } else if (s_ini->socks_version == PROXY_PROTO_SOCKS_V4) {
                     printl(LOG_VERB, "Initiate Socks4 protocol: request: [%s] -> [%s]",
-                        inet2str(&s_ini->socks_server, suf), inet2str(&daddr, buf));
+                        inet2str(&s_ini->socks_server, suf), inet2str(&daddr_u.ip_addr, buf));
+
+                    /* TODO: Use daddr_u only! */
                     if (socks4_client_request(ssock, SOCKS4_CMD_TCPCONNECT,
-                            (struct sockaddr_in *)&daddr, s_ini->socks_user) != SOCKS4_REPLY_OK) {
+                            (struct sockaddr_in *)&daddr_u.ip_addr, s_ini->socks_user) != SOCKS4_REPLY_OK) {
                 
                         printl(LOG_WARN, "Socks4 server returned an error");
                         close(csock);